+++ pom.xml
@@ -15,6 +15,8 @@
 	<description>Practice project for Spring Boot</description>
 	<properties>
 		<java.version>11</java.version>
+		<sonar.organization>nagarwala</sonar.organization>
+		<sonar.host.url>https://sonarcloud.io</sonar.host.url>
 	</properties>
 	<dependencies>
 		<dependency>
@@ -107,6 +109,7 @@
 					</execution>
 				</executions>
 			</plugin>
+
 		</plugins>
 	</build>
