--- src/main/java/com/springBoot/GreenCommute/GreenCommuteApplication.java
@@ -1,13 +1,0 @@
-package com.springBoot.GreenCommute;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class GreenCommuteApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(GreenCommuteApplication.class, args);
-	}
-
-}
