--- src/main/java/com/springBoot/GreenCommute/dto/JobDto.java
@@ -1,23 +1,0 @@
-package com.springBoot.GreenCommute.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class JobDto {
-
-    private int jobId;
-
-    private String jobTitle;
-
-    private String jobLocation;
-
-    private List<String> commuteOptions;
-
-}
-
