--- src/main/java/com/springBoot/GreenCommute/dto/UserDto.java
@@ -1,16 +1,0 @@
-package com.springBoot.GreenCommute.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class UserDto {
-
-    private int userId;
-
-    private String userName;
-}
-
