--- src/main/java/com/springBoot/GreenCommute/entities/Skill.java
@@ -1,31 +1,0 @@
-package com.springBoot.GreenCommute.entities;
-
-import com.fasterxml.jackson.annotation.JsonBackReference;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.util.List;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter @Setter
-@Entity
-@Table(name = "skill")
-public class Skill {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "skill_id")
-    private int skillId;
-
-    @Column(name = "skill_name")
-    private String skillName;
-
-    @ManyToMany(mappedBy = "skillList",cascade = CascadeType.ALL)
-    @JsonBackReference
-    private List<Job> jobList;
-}
-
