--- src/main/java/com/springBoot/GreenCommute/entities/User.java
@@ -1,30 +1,0 @@
-package com.springBoot.GreenCommute.entities;
-
-import com.fasterxml.jackson.annotation.JsonManagedReference;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.util.List;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter @Setter
-@Entity
-@Table(name = "user")
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private int userId;
-
-    @Column(name = "user_name")
-    private String userName;
-
-    @OneToMany(mappedBy = "user",cascade = {CascadeType.DETACH, CascadeType.MERGE, CascadeType.PERSIST, CascadeType.REFRESH})
-    @JsonManagedReference
-    private List<SavedJob> savedJobList;
-}
