--- src/main/java/com/springBoot/GreenCommute/exception/DataExceptionHandler.java
@@ -1,29 +1,0 @@
-package com.springBoot.GreenCommute.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class DataExceptionHandler {
-
-    @ExceptionHandler
-    public ResponseEntity<ErrorResponse> handleException(DataNotFoundException exc){
-        ErrorResponse errorResponse = new ErrorResponse();
-        errorResponse.setStatus(HttpStatus.NOT_FOUND.value());
-        errorResponse.setMessage(exc.getMessage());
-        errorResponse.setTimeStamp(System.currentTimeMillis());
-        return new ResponseEntity<>(errorResponse,HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<ErrorResponse> handleBadRequest(Exception exc){
-        ErrorResponse errorResponse = new ErrorResponse();
-        errorResponse.setStatus(HttpStatus.BAD_REQUEST.value());
-        errorResponse.setMessage(exc.getMessage());
-        errorResponse.setTimeStamp(System.currentTimeMillis());
-        return new ResponseEntity<>(errorResponse,HttpStatus.BAD_REQUEST);
-    }
-
-}
