--- src/main/java/com/springBoot/GreenCommute/exception/DataNotFoundException.java
@@ -1,15 +1,0 @@
-package com.springBoot.GreenCommute.exception;
-
-public class DataNotFoundException extends RuntimeException{
-    public DataNotFoundException(String message) {
-        super(message);
-    }
-
-    public DataNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public DataNotFoundException(Throwable cause) {
-        super(cause);
-    }
-}
