--- src/main/java/com/springBoot/GreenCommute/exception/ErrorResponse.java
@@ -1,17 +1,0 @@
-package com.springBoot.GreenCommute.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter @Setter
-public class ErrorResponse {
-
-    private int status;
-    private String message;
-    private long timeStamp;
-
-}
