--- src/main/java/com/springBoot/GreenCommute/repositories/SavedJobRepository.java
@@ -1,14 +1,0 @@
-package com.springBoot.GreenCommute.repositories;
-
-import com.springBoot.GreenCommute.entities.SavedJob;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface SavedJobRepository extends JpaRepository<SavedJob,Integer> {
-
-    @Query(value = "select * from jobs_saved where user_id=?1 and jobs_id=?2",nativeQuery = true)
-    SavedJob findByUserAndJobId(int userId, int jobId);
-}
-
