--- src/main/java/com/springBoot/GreenCommute/repositories/UserRepository.java
@@ -1,9 +1,0 @@
-package com.springBoot.GreenCommute.repositories;
-
-import com.springBoot.GreenCommute.entities.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User,Integer> {
-}
