--- src/main/java/com/springBoot/GreenCommute/service/SavedJobService.java
@@ -1,16 +1,0 @@
-package com.springBoot.GreenCommute.service;
-
-import com.springBoot.GreenCommute.dto.JobDto;
-import com.springBoot.GreenCommute.entities.Job;
-
-import java.util.List;
-
-public interface SavedJobService {
-
-    void addToSavedJob(int userId, int jobId);
-
-    void deleteFromSavedJob(int userId, int jobId);
-
-    List<JobDto> getSavedJobsForUser(int userId);
-}
-
