--- src/main/java/com/springBoot/GreenCommute/service/UserServiceImpl.java
@@ -1,36 +1,0 @@
-package com.springBoot.GreenCommute.service;
-
-import com.springBoot.GreenCommute.entities.User;
-import com.springBoot.GreenCommute.repositories.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.Optional;
-
-@Service
-public class UserServiceImpl implements UserService{
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @Override
-    public List<User> getUserList() {
-        return userRepository.findAll();
-    }
-
-    @Override
-    public Optional<User> getUserById(int userId) {
-        return userRepository.findById(userId);
-    }
-
-    @Override
-    public void addUser(User user) {
-        userRepository.save(user);
-    }
-
-    @Override
-    public void deleteUserById(int userId) {
-        userRepository.deleteById(userId);
-    }
-}
