+++ src/main/java/com/springboot/GreenCommute/GreenCommuteApplication.java
@@ -1,0 +1,13 @@
+package com.springboot.GreenCommute;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GreenCommuteApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(GreenCommuteApplication.class, args);
+	}
+
+}
