+++ src/main/java/com/springboot/GreenCommute/dto/UserDto.java
@@ -1,0 +1,16 @@
+package com.springboot.GreenCommute.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class UserDto {
+
+    private int userId;
+
+    private String userName;
+}
+
