+++ src/main/java/com/springboot/GreenCommute/exception/DataNotFoundException.java
@@ -1,0 +1,15 @@
+package com.springboot.GreenCommute.exception;
+
+public class DataNotFoundException extends RuntimeException{
+    public DataNotFoundException(String message) {
+        super(message);
+    }
+
+    public DataNotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public DataNotFoundException(Throwable cause) {
+        super(cause);
+    }
+}
