+++ src/main/java/com/springboot/GreenCommute/exception/ErrorResponse.java
@@ -1,0 +1,17 @@
+package com.springboot.GreenCommute.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter @Setter
+public class ErrorResponse {
+
+    private int status;
+    private String message;
+    private long timeStamp;
+
+}
