+++ src/main/java/com/springboot/GreenCommute/repositories/UserRepository.java
@@ -1,0 +1,9 @@
+package com.springboot.GreenCommute.repositories;
+
+import com.springboot.GreenCommute.entities.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User,Integer> {
+}
