+++ src/main/java/com/springboot/GreenCommute/service/UserService.java
@@ -1,0 +1,18 @@
+package com.springboot.GreenCommute.service;
+
+import com.springboot.GreenCommute.entities.User;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface UserService {
+
+    public List<User> getUserList();
+
+    public Optional<User> getUserById(int userId);
+
+    public void addUser(User user);
+
+    public void deleteUserById(int userId);
+}
+
