--- src/test/java/com/springBoot/GreenCommute/GreenCommuteApplicationTests.java
@@ -1,13 +1,0 @@
-package com.springBoot.GreenCommute;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GreenCommuteApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
