+++ src/test/java/com/springboot/GreenCommute/GreenCommuteApplicationTests.java
@@ -1,0 +1,13 @@
+package com.springboot.GreenCommute;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GreenCommuteApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
